Add MLSCiphertextContent
