+++ include/mls/messages.h
@@ -539,8 +539,8 @@ struct MLSPlaintext
   MLSPlaintext(bytes group_id,
                epoch_t epoch,
                Sender sender,
-               ContentType content_type,
                bytes authenticated_data,
+               ContentType content_type,
                const bytes& content);
 
   // Constructors for encrypting
@@ -592,6 +592,47 @@ private:
   // Not part of the struct, an indicator of whether this MLSPlaintext was
   // constructed from an MLSCiphertext
   bool decrypted;
+};
+
+// struct {
+//     select (MLSCiphertext.content_type) {
+//         case application:
+//           opaque application_data<0..2^32-1>;
+//         case proposal:
+//           Proposal proposal;
+//         case commit:
+//           Commit commit;
+//     }
+//     opaque signature<0..2^16-1>;
+//     optional<MAC> confirmation_tag;
+//     opaque padding<0..2^16-1>;
+// } MLSCiphertextContent;
+struct MLSCiphertextContent
+{
+  const var::variant<ApplicationData, Proposal, Commit>& content;
+  const bytes& signature;
+  const std::optional<MAC>& confirmation_tag;
+  const bytes& padding;
+};
+
+tls::ostream&
+operator<<(tls::ostream& str, const MLSCiphertextContent& ct);
+
+// struct {
+//     opaque group_id<0..255>;
+//     uint64 epoch;
+//     ContentType content_type;
+//     opaque authenticated_data<0..2^32-1>;
+// } MLSCiphertextContentAAD;
+struct MLSCiphertextContentAAD
+{
+  const bytes& group_id;
+  const epoch_t epoch;
+  const ContentType content_type;
+  const bytes& authenticated_data;
+
+  TLS_SERIALIZABLE(group_id, epoch, content_type, authenticated_data)
+  TLS_TRAITS(tls::vector<1>, tls::pass, tls::pass, tls::vector<4>)
 };
 
 // struct {
